projects/etl_job_rerun/requirements.txt
Flask==3.0.0
boto3==1.34.0
botocore==1.34.0
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3

# Optional: For Snowflake cleanup hooks
snowflake-connector-python==3.6.0

# Optional: For additional data processing
openpyxl==3.1.2

# Note: Also required in project root:
# - env_config.py (AWS profile to bucket mappings)